`vms/platformvm`: Remove `MempoolTxVerifier` (#2362)
