+++ vms/platformvm/txs/executor/standard_tx_executor_test.go
@@ -179,7 +179,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 		setup                func(*environment)
 		AP3Time              time.Time
 		expectedExecutionErr error
-		expectedMempoolErr   error
 	}
 
 	tests := []test{
@@ -194,7 +193,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                nil,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrPeriodMismatch,
-			expectedMempoolErr:   ErrPeriodMismatch,
 		},
 		{
 			description:          fmt.Sprintf("delegator should not be added more than (%s) in the future", MaxFutureStartTime),
@@ -207,7 +205,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                nil,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrFutureStakeTime,
-			expectedMempoolErr:   nil,
 		},
 		{
 			description:          "validator not in the current or pending validator sets",
@@ -220,7 +217,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                nil,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: database.ErrNotFound,
-			expectedMempoolErr:   database.ErrNotFound,
 		},
 		{
 			description:          "delegator starts before validator",
@@ -233,7 +229,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                addMinStakeValidator,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrPeriodMismatch,
-			expectedMempoolErr:   ErrPeriodMismatch,
 		},
 		{
 			description:          "delegator stops before validator",
@@ -246,7 +241,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                addMinStakeValidator,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrPeriodMismatch,
-			expectedMempoolErr:   ErrPeriodMismatch,
 		},
 		{
 			description:          "valid",
@@ -259,7 +253,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                addMinStakeValidator,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: nil,
-			expectedMempoolErr:   nil,
 		},
 		{
 			description:          "starts delegating at current timestamp",
@@ -272,7 +265,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                nil,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrTimestampNotBeforeStartTime,
-			expectedMempoolErr:   ErrTimestampNotBeforeStartTime,
 		},
 		{
 			description:   "tx fee paying key has no funds",
@@ -297,7 +289,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			},
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrFlowCheckFailed,
-			expectedMempoolErr:   ErrFlowCheckFailed,
 		},
 		{
 			description:          "over delegation before AP3",
@@ -310,7 +301,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                addMaxStakeValidator,
 			AP3Time:              defaultValidateEndTime,
 			expectedExecutionErr: nil,
-			expectedMempoolErr:   nil,
 		},
 		{
 			description:          "over delegation after AP3",
@@ -323,7 +313,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			setup:                addMaxStakeValidator,
 			AP3Time:              defaultGenesisTime,
 			expectedExecutionErr: ErrOverDelegated,
-			expectedMempoolErr:   ErrOverDelegated,
 		},
 	}
 
@@ -363,15 +352,6 @@ func TestStandardTxExecutorAddDelegator(t *testing.T) {
 			}
 			err = tx.Unsigned.Visit(&executor)
 			require.ErrorIs(err, tt.expectedExecutionErr)
-
-			mempoolExecutor := MempoolTxVerifier{
-				Backend:       &freshTH.backend,
-				ParentID:      lastAcceptedID,
-				StateVersions: freshTH,
-				Tx:            tx,
-			}
-			err = tx.Unsigned.Visit(&mempoolExecutor)
-			require.ErrorIs(err, tt.expectedMempoolErr)
 		})
 	}
 }
